Introduce module for firebase references
