+++ TemplateApplication/TemplateApplicationStandard.swift
@@ -26,54 +26,11 @@ actor TemplateApplicationStandard: Standard,
                                    HealthKitConstraint,
                                    OnboardingConstraint,
                                    AccountNotifyConstraint {
-    enum TemplateApplicationStandardError: Error {
-        case userNotAuthenticatedYet
-    }
-
-    static var userCollection: CollectionReference {
-        Firestore.firestore().collection("users")
-    }
-
     @Application(\.logger) private var logger
 
-    @Dependency(Account.self) private var account: Account? // optional, as Firebase might be disabled
-    @Dependency(FirebaseAccountService.self) private var accountService: FirebaseAccountService?
-
-    
-    private var userDocumentReference: DocumentReference {
-        get async throws {
-            guard let details = await account?.details else {
-                throw TemplateApplicationStandardError.userNotAuthenticatedYet
-            }
-
-            return userDocumentReference(for: details.accountId)
-        }
-    }
-    
-    private var userBucketReference: StorageReference {
-        get async throws {
-            guard let details = await account?.details else {
-                throw TemplateApplicationStandardError.userNotAuthenticatedYet
-            }
-
-            return Storage.storage().reference().child("users/\(details.accountId)")
-        }
-    }
-
+    @Dependency(FireConfiguration.self) private var configuration
 
     init() {}
-
-
-    private func userDocumentReference(for accountId: String) -> DocumentReference {
-        Self.userCollection.document(accountId)
-    }
-
-
-    nonisolated func configure() {
-        Task {
-            await setupTestAccount()
-        }
-    }
 
 
     func add(sample: HKSample) async {
@@ -113,7 +70,7 @@ actor TemplateApplicationStandard: Standard,
         }
         
         do {
-            try await userDocumentReference
+            try await configuration.userDocumentReference
                 .collection("QuestionnaireResponse") // Add all HealthKit sources in a /QuestionnaireResponse collection.
                 .document(id) // Set the document identifier to the id of the response.
                 .setData(from: response)
@@ -124,7 +81,7 @@ actor TemplateApplicationStandard: Standard,
     
     
     private func healthKitDocument(id uuid: UUID) async throws -> DocumentReference {
-        try await userDocumentReference
+        try await configuration.userDocumentReference
             .collection("HealthKit") // Add all HealthKit sources in a /HealthKit collection.
             .document(uuid.uuidString) // Set the document identifier to the UUID of the document.
     }
@@ -132,7 +89,7 @@ actor TemplateApplicationStandard: Standard,
     func respondToEvent(_ event: AccountNotifications.Event) async {
         if case let .deletingAccount(accountId) = event {
             do {
-                try await userDocumentReference(for: accountId).delete()
+                try await configuration.userDocumentReference(for: accountId).delete()
             } catch {
                 logger.error("Could not delete user document: \(error)")
             }
@@ -168,41 +125,11 @@ actor TemplateApplicationStandard: Standard,
 
             let metadata = StorageMetadata()
             metadata.contentType = "application/pdf"
-            _ = try await userBucketReference
+            _ = try await configuration.userBucketReference
                 .child("consent/\(dateString).pdf")
                 .putDataAsync(consentData, metadata: metadata) { @Sendable _ in }
         } catch {
             await logger.error("Could not store consent form: \(error)")
-        }
-    }
-
-    private func setupTestAccount() async {
-        guard let accountService, FeatureFlags.setupTestAccount else {
-            return
-        }
-
-        do {
-            try await accountService.login(userId: "lelandstanford@stanford.edu", password: "StanfordRocks!")
-            return
-        } catch {
-            guard let accountError = error as? FirebaseAccountError,
-                  case .invalidCredentials = accountError else {
-                logger.error("Failed to login into test account: \(error)")
-                return
-            }
-        }
-
-        // account doesn't exist yet, signup
-        var details = AccountDetails()
-        details.userId = "lelandstanford@stanford.edu"
-        details.password = "StanfordRocks!"
-        details.name = PersonNameComponents(givenName: "Leland", familyName: "Stanford")
-        details.genderIdentity = .male
-
-        do {
-            try await accountService.signUp(with: details)
-        } catch {
-            logger.error("Failed to setup test account: \(error)")
         }
     }
 }
